// projects/codx-ac/src/lib/vouchers/purchaseinvoices/interfaces/IPurchaseInvoice.inteface.ts
export interface IPurchaseInvoice {
    recID: string;
    journalNo: string;
    journalType: string;
    subType: string;
    voucherNo: string;
    voucherDate: string;
    invoiceType: string;
    invoiceForm: string;
    invoiceSeriNo: string;
    invoiceNo: string;
    invoiceDate: string | null;
    invoiceDueDate: string | null;
    objectType: string;
    objectID: string;
    objectName: string;
    address: string;
    taxCode: string;
    partnerID: string;
    currencyID: string;
    exchangeRate: number;
    multi: boolean;
    memo: string;
    memo2: string;
    status: string;
    totalAmt: number;
    totalDiscPct: number;
    totalDiscAmt: number;
    taxExchRate: number;
    taxMulti: boolean;
    pmtMethodID: string;
    pmtTermID: string;
    delModeID: string;
    warehouseID: string;
    reasonID: string;
    buyer: string;
    orderPoolID: string;
    orderNo: string;
    contractID: string;
    projectID: string;
    diM1: string;
    diM2: string;
    settlement: string;
    settleVouchers: string;
    cancelled: string;
    cancelReason: string;
    refType: string;
    refNo: string;
    refID: string | null;
    bLNo: string;
    bLDate: string | null;
    printed: number | null;
    autoCreated: boolean;
    postedLayer: string;
    postedDate: string | null;
    singleEntry: string;
    owner: string;
    buid: string;
    createdOn: string;
    createdBy: string;
    modifiedOn: string | null;
    modifiedBy: string;

    unbounds?: any;
}